+++ source/components/HOC/withTheme.js
@@ -3,6 +3,7 @@ import React from 'react';
 import type { ComponentType, Ref } from 'react';
 import forwardRef from 'create-react-ref/lib/forwardRef';
 import { ThemeContext } from './ThemeContext';
+import { getDisplayName } from '../../utils/props';
 
 // withTheme is a HOC that takes a Component as a parameter
 // and returns that Component wrapped within ThemeContext.Consumer.
@@ -26,5 +27,7 @@ export const withTheme = (Component: ComponentType<any>) => {
     ));
   }
 
+  // create a new displayName for the wrapped component
+  WrappedComponent.displayName = `withTheme(${getDisplayName(Component)})`;
   return WrappedComponent;
 };