+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ## Features
 
+- Adds displayName static property to all components
+  [PR 89](https://github.com/input-output-hk/react-polymorph/pull/89)
+
 - Adds InfiniteScroll component, InfiniteScrollSkin to simple skins, and SimpleInfiniteScroll to SimpleTheme.
   [PR 88](https://github.com/input-output-hk/react-polymorph/pull/88)